// types.ts
export enum UserRole {
  Student = 'Student',
  Teacher = 'Teacher',
  Admin = 'Admin',
}

export enum ClassPreference {
  Online = 'Online',
  Offline = 'Offline',
  Hybrid = 'Hybrid', // Added for teachers
}

export enum Sex {
    Male = 'Male',
    Female = 'Female',
    Other = 'Other',
}

export enum EmploymentType {
    PartTime = 'Part-time',
    FullTime = 'Full-time',
}

export enum Grade {
    Grade1 = 'Grade 1',
    Grade2 = 'Grade 2',
    Grade3 = 'Grade 3',
}

export interface Document {
  name: string;
  mimeType: string;
  data: string; // base64
}

export enum UserStatus {
    Active = 'Active',
    Inactive = 'Inactive',
    OnHold = 'On Hold',
    Graduated = 'Graduated',
}

export interface Location {
  id: string;
  name: string;
  address: string;
}


export interface User {
  id: string;
  name: string;
  email: string;
  password?: string; // In a real app, this should be a secure hash. Optional on client.
  role: UserRole;
  classPreference?: ClassPreference; 
  
  // Common fields
  photoUrl?: string; // base64 or URL
  dob?: string; // ISO string date
  sex?: Sex;
  contactNumber?: string;
  alternateContactNumber?: string;
  address?: string;
  schedules?: { course: string, timing: string, teacherId?: string }[];
  documents?: Document[];
  dateOfJoining?: string;
  country?: string;
  state?: string;
  city?: string;
  postalCode?: string;
  timezone?: string;
  preferredTimings?: string[];
  status?: UserStatus;
  locationId?: string; // For offline preference
  location?: Location; // Populated field

  // Student specific
  courses?: string[];
  fatherName?: string;
  standard?: string;
  schoolName?: string;
  grade?: Grade;
  notes?: string;
  
  // Teacher specific
  courseExpertise?: string[];
  educationalQualifications?: string;
  employmentType?: EmploymentType;
  yearsOfExperience?: number;
  availableTimeSlots?: string[];

  // Soft delete fields
  isDeleted?: boolean;
  deletedAt?: string;
}